# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing user registration, logins, and credential storage.",
        "title": "Gin Password Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Verifies credentials and generates a session JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Authenticate user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.AuthRes"
                        }
                    },
                    "400": {
                        "description": "Invalid input credentials",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new user in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.AuthRes"
                        }
                    },
                    "400": {
                        "description": "Could not create user",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    }
                }
            }
        },
        "/passwordcrud/all": {
            "get": {
                "description": "Retrieve all password credential records belonging to the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passwords"
                ],
                "summary": "List all password records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT_TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.PasswordRes"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    }
                }
            }
        },
        "/passwordcrud/create": {
            "post": {
                "description": "Create a password credential record for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passwords"
                ],
                "summary": "Create password record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT_TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Password payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.PasswordCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.PasswordRes"
                        }
                    },
                    "400": {
                        "description": "Could not create password",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    }
                }
            }
        },
        "/passwordcrud/delete/{id}": {
            "delete": {
                "description": "Deletes a specific password record by ID, verifying that it belongs to the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passwords"
                ],
                "summary": "Delete a password record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT_TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "400": {
                        "description": "Could not delete password",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "404": {
                        "description": "Password not found or not authorized to delete",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    }
                }
            }
        },
        "/passwordcrud/update/{id}": {
            "patch": {
                "description": "Partially update a specific password record's title, email, or password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passwords"
                ],
                "summary": "Update a password record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT_TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.PasswordUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "404": {
                        "description": "Password record not found or unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    }
                }
            }
        },
        "/protected/getuser": {
            "get": {
                "description": "Retrieve the logged-in user profile info using the user ID loaded from the JWT claims.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT_TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserProfileRes"
                        }
                    },
                    "400": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.AuthRes": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "dtos.MessageRes": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "dtos.PasswordCreateReq": {
            "type": "object",
            "required": [
                "email",
                "password",
                "title"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "dtos.PasswordRes": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dtos.PasswordUpdateReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "dtos.UserLoginReq": {
            "type": "object",
            "required": [
                "email",
                "mpin",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "mpin": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 6
                }
            }
        },
        "dtos.UserProfileRes": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dtos.UserRegisterReq": {
            "type": "object",
            "required": [
                "email",
                "mpin",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "mpin": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 6
                }
            }
        }
    }
}